Remove old function for testing email templates.
This was written before we had the code to allow organizers to preview
email templates before sending them. Now we should be able to create a
scenario that allows us to create an internship round at the right time,
log in as an organizer, and double check the email templates.