Use snippet/eligibility_prompts instead of things it includes
The eligibility_prompts snippet may display quite a few different
messages, and in some cases we know we only want to display a specific
one of those messages. However, keeping the individual specific messages
in separate snippets has been difficult to maintain, so we want to merge
them all into the main eligibility_prompts.html snippet.

This commit replaces includes of specific snippets with includes of the
general eligibility_prompts.html instead.

I'm assuming that nobody can be both an approved coordinator and an
approved applicant at the same time, which means the general snippet
will only show things which are applicable to one or the other.

The project_contributions.html template is only rendered for approved
applicants, so the general snippet will always either show
contribute_prompt or apply_prompt, depending on whether the applicant
has already made contributions in this round. This page previously used
the same logic except it didn't display anything in the
contribute_prompt case. Adding that text here may be redundant but Sage
will review it for clarity later.

The dashboard/coordinator_reminder.html template should render exactly
the same except it might also prompt the coordinator to agree to the
code of conduct if they haven't already.