Make Role.today consistent with date used to find current_round
I couldn't do this everywhere without a larger refactoring, but where I
could, this means anything using the Role doesn't have a race condition
from picking a different value of `today` than was used to decide which
round is current. It also reduces the number of calls to datetime.now.