Rename role.needs_application to is_potential_applicant
Make all instances of changing the eligibility breadcrumb link and text
to always point to the eligibility page. If someone filled out an
initial application and they try to do so again, they will get
redirected to their initial application results. So it's fine to always
direct them to that page.