Include doctests in our Django test suite
Sometimes a function is simple enough that it makes sense to include its
unit tests directly in its documentation, so they can double as examples
of how to use the function. This commit makes `./manage.py test` run
those doctests.

At the moment we only have doctests in home.models so I only added that
module to the test suite. If you add doctests to other modules you'll
need to list those modules in test_docs.py as well.