Port the Outreachy promotion page over to Django.
This page has always been a pain because it contains dates that change.

Update all the dates we currently have model fields for. The only dates
we don't have fields for is the #OutreachyChat date.

Update the sample email to reflect the new initial application and
contribution period difference. Update the sample tweets to use date
fields and to fit into 280 character limitations.

Update the links to use url look ups when possible. The /apply/ page is
not currently a Django page, so we can't update that.