Replace several {% url %}/reverses with model methods
The following view names no longer appear anywhere outside urls.py and
models.py:

- applicant-review-detail
- application-action (for both initial and final applications)
- contributions
- coordinatorapproval-action
- mentorapproval-action
- project-applicants
- select-intern

In many cases we have an object available which fully determines the
keyword arguments needed to construct some URL. For example, if we have
a Participation and we need a URL that only requires a round_slug and a
community_slug, then a method on Participation can construct that URL.

In this commit I have replaced all remaining {%url%} template tags and
calls to reverse for view names which already had at least one model
method somewhere.

In some cases the existing model method wasn't applicable in some of the
places where a tag could use it. In those cases I introduced a similar
method on a different object that the call site did have access to.