Introduce model-specific reverse() helpers
We have a lot of calls to reverse() with very similar arguments, because
for example, any view which takes a project needs the round_slug,
community_slug, and project_slug, in addition to any view-specific
arguments.

So on several key models, I've added a new reverse() method which calls
the real reverse() but adds all the appropriate arguments for that
model. In the above example, Project.reverse adds the three slugs to
whatever other keyword arguments the caller provides.