Remove buggy test.
The SyntaxTestCase tests whether a series of URLs are reachable. The
problem is that it creates a intern visitor type by calling into
InternSelectionFactory(). That will create a new RoundPage with
internannounce set to today's date.

The problem is that coordinators shouldn't be allowed to submit or
withdraw from the round after interns are announced. So this test was
buggy to begin with - those pages should have returned a 404. That means
the tests for submitting projects probably have the same bug.

For now, comment out the code that was failing because the view was
fixed to properly return a 404. TODO: add more tests that test community
approval, rejection, and withdrawal.