Stop using transaction.on_commit, for easier testing
I wrote ApprovalStatusAction to use transaction.on_commit so that emails
would only get sent after we're certain the corresponding database
transaction has safely hit the disk.

This was just a precaution. The transaction.on_commit call is the very
last thing we do before returning from the view, so deferring it only
moves it past response middleware, which hopefully we can rely on.

Turns out that Django's TestCase causes transaction.on_commit to never
fire in tests, because it works by always rolling back the transaction
at the end of each test.

https://docs.djangoproject.com/en/1.11/topics/db/transactions/#use-in-tests

We could switch to TransactionTestCase for anything that needs to test
email sending in views using this mixin... but that gives us the choice
of test suite problems with either correctness or performance.

So:

- on_commit isn't protecting us from anything we're particularly
  concerned about.
- By removing on_commit we can write unit tests that will help us catch
  the problems on_commit would have mitigated.

Therefore I'm just going to stop using it.