Use action-form snippet for non-submit participation actions
ParticipationAction is unusual because its form_class is an
InlineFormSet, not a ModelForm. So rendering it to HTML while
customizing the details uses a different loop structure, which doesn't
work if you get handed a ModelForm instead. And when the action is not
"submit", the template is handed a ModelForm, so it didn't generate
sensible HTML for anything in those cases.