Improve organizer workflow for approving community participation.
Avoid a double approval screen when approving a community. Only ask for
a double check when rejecting a community (since we need a reason for
denying a community).

Layout approve and reject buttons with Bootstrap's grid system. This
will cause the reject button to be off to the far right when viewing on
a desktop system. Add margins around the buttons to allow for spacing in
between the buttons when viewing on a mobile device.