Fix bug in community CFP page and add test to catch bug in the future.
The community CFP page should only display a prompt to submit a new
community from the time the call for communities goes out
(RoundPage.pingnew) and before the sign up for new communities is closed
(RoundPage.lateorgs).

Add a test to the early, on-time, and late community sign up to ensure
the 'Submit new community' button is available. There's no tests yet to
submit a new community, but this will help some.