Fix usage cases for initial or final application deadline in prompts.
When the final application deadline had passed, we were displaying that
prompt twice on the project selection page. We might also be incorrectly
telling an applicant they did not submit a final application.

The template home/templates/home/snippet/deadline_passed.html was used
to indicate both initial application deadline had passed and the
contribution deadline had passed. Change the templates that included
that snippet to be clear about which deadline it was talking about.