Work around wontfix'd bug pypa/pipenv#1050
Pipenv developers argue that locking Python to a specific minor release
like "3.6" is good since it documents what we're testing this project
with. There's some truth to that, although leaving off the patch release
("3.6.3") seems like a strange omission in that case. For that matter,
Pipenv doesn't provide any way to pin any non-Python dependencies or
declare which platform(s) the project is tested on, so the whole
argument is pretty weak.

We need contributors to have as easy of a setup process as possible, and
that includes using the version of Python they have available. This
Pipenv "feature" has caused confusion for new contributors; its cost is
very much not justified by the minor benefits it gives to
reproducibility.

Apparently the only way to make Pipenv do what we need is to remove the
version declaration entirely, so that's what I'm doing. We'll just have
to be aware that if we get bug reports that we can't reproduce, it may
be because someone is using a too-old version of Python.