2019-12 round: Add blog post for picking projects
This blog post represents a sample of what we could do if we had a
tagging system for projects. This was a tedious effort done by hand, and
prone to errors in misunderstanding the project descriptions or
accidentally leaving out projects. But it also shows a sample of
normalized tags we could add that mentors could self-tag their project
as.