Fix bug that sends intern selection emails to rejected, pending, and withdrawn mentors.
After a mentor selects an intern, we want to email co-mentors for the
project to tell them to sign the mentor agreement.

The previous code was getting all co-mentors for a project by using
project.mentorapproval_set.all(). However, that query includes mentors
who have been rejected, pending, or have withdrawn from mentorship. Fix
the query to only get approved co-mentors by using
project.mentorapproval_set.approved() instead.

Write a test to catch this mistake if it happens again in the future,
and also ensure that the correct links are sent in the email to the
co-mentor. This also adds a new scenario to home/scenarios.py for when
the contribution period closes but interns haven't been selected yet.

Fixes #359