Add blog and chat dates to internship guide
Also move some details from the welcome email to the internship guide.
Add short blog prompt titles for each week. I didn't want to move the
full text of each blog prompt over to the internship guide yet. It needs
to be carefully scrubbed so that links to the Zulip chat or the contact
details for our resume reviewer don't leak into it.