Add link to internship guide and applicant guide to apply page
People often don't look in the menu for information. Most people who are
looking for information about applying to Outreachy eventually end up on
the eligibility page. (And then they send us email asking when the
application period will be open.)

Link to the applicant guide and internship guide on this page. The page
at /apply/ is under Wagtail CMS control, so links will have to be added
to the guides manually through the website.