Fix bug in displaying intern pronouns.
Some interns have a flag set to display their pronouns publicly on the
alums page, but don't have the flag set to display their pronouns to
their mentors. That could be because they want to hide their gender
during the contribution process in order to not bias mentors. But once
interns are announced, I think it makes sense to use the pronouns listed
on the alums page.