Draft work for adding warnings about time commitment changes.
The goal of this work is to catch applicants who accept a full-time job
in between filling out the initial application and the intern selection.
Some people continue to contribute because they want to work full-time
and do the Outreachy internship full-time in their evenings and
weekends. If applicants have a full-time job (or a job over 20 hours a
week), they aren't eligible for the internship.

Fix this by adding notes to the email template for the initial
application acceptance emails. Add fields to the final applications to
give applicants a chance to update their time commitment information.

Todo: Change field names to time_correct and time_comments.
I'm getting mixed up when I try to remember if it's
"time_commitments_correct" or "free_time_correct".

Add tests to make sure these warnings show up on the organizer dash.
Coordinators and mentors should not be able to see the time commitment
details or updates (as they may be a privacy violation in some cases).

Maybe make a "this applicant has updated time commitments, and may be
ineligible. please check with outreachy organizers" comment on the
application? But how do they know we've reviewed the updated time
commitments and it's okay? time_changes_approved ?

Update the checkbox on the final application form to use a yes/no
radial. While we're at it, the layout of the final application form
doesn't match the layout used in the initial application form. It's
really not very readable.