All applications show as pending until contribution period opens.
We know that some applicants fill out a second initial application after
they get rejected for time. The Outreachy organizers also find it hard
to deal with the number of applicants that ask to get their application
reset due to failing a 'general' question, e.g. are they eligible to
work 40 hours a week.

This round, we know there will be some applications we don't get to,
partially because of these duplicate applications, and partially because
demand for our program has grown.

That means some applications won't be reviewed by organizers and will be
pending when the application period opens.

Make all applications show up as being under review until the
contribution period opens. After that, show people why they were
rejected, with a generic message for people whose application we didn't
get a chance to review.

Modify the tests to check that all applications are marked as under
review. Add test to check correct status for pending applications after
the contribution period opens.