Fix bug that was showing pending applications as still pending after the contribution period opens
Once the contribution period opens, any application we couldn't get to
review, or where people didn't answer our questions needs to show up as
rejected.

There was a bug in the template that incorrectly looked for
RoundPage.contribution_opens instead of RoundPage.contributions_open.

Because we didn't have that many applicants who were pending in previous
rounds, we didn't notice this bug.