chore: easier to read contributions
The use of `<pre>` tags for the description of Contributions on Review
Project Applicants page are hard to read. Given that most aplicants
don't format the description the text is placed on a single line with a
horizontal scrolling.

This changes the `<pre>` tag to `<blockquote>` and styles it by adding
the left border in the theme colour. The `<blockquote>` word wraps and
preserves any whitespace and newlines with `white-space: pre-wrap` so
both the pre-formatting readability can be achieved.