Feedback: Update forms and put in temp Covid-19 policy.
This is a complete rework of the intern and mentor feedback forms to
make them clearer, and also putting in changes to support our temporary
Covid-19 internship weekly time commitment policies.

Sorry future self for combining these two commits. :-/

New Fields
---

New fields in the intern feedback forms:

 - Time comments. This allows them to tell us why they haven't been
working 40 hours a week

 - Opt-in information sharing with the community coordinator. The
default is opt-out.

New field in the mentor feedback forms:

 - Mentor report. This is a self-evaluation that mentors fill out. The
information is shared with Outreachy organizers and coordinators.

Change the test suite to handle the new fields.

Privacy Changes
---

There is a new privacy change to the mentor feedback forms. Coordinators
will be able to see all the answers mentors provide.

If the intern opts-in, we share their comments about how their mentor
has been supporting them, and metrics about their mentor support with
their community coordinator. This was requested by some coordinators in
order to get more intern sponsorship and support mentors' request for
employers to support their volunteer time with Outreachy interns.

Add a warning that if the intern opts-in to sharing comments with a
coordinator, and their mentor is a coordinator, they will be sharing the
information directly with the mentor. Add a note as to who is currently
a coordinator and who is a mentor for that intern.

Note: This doesn't cover the case where a mentor is later approved to be
a coordinator.

Covid-19 policies
---

Change various form names and help texts to remove references to
"full-time" or "40 hours a week". Add table that describes our temporary
minimum time commitment policies due to Covid-19.

Clean-up
---

 - Rearrange forms to have sections that make sense to the person
filling them out. Have sections for self-evaluation and for evaluating
the other person (i.e. intern or mentor).

 - Clarify when mentors should ask for an extension or a termination

 - more paragraphs to make it easier to read