Email: Add invitation email for informal chat interviewees
After the careers chat in week 10, interns find out that networking is
very important to their career. In week 12, we talk about how to do an
informal chat to expand your network.

That means we need people for interns to interview! A month before week
12 (so during week 8), we ping coordinators, mentors, alums, and
sponsors to ask them if they want to participate in informal chats.

We have some very specific rules about who can participate in an
informal chat with an Outreachy intern.

Interviewees must be:

 - a past Outreachy intern (alum)
 - a past or current Outreachy mentor
 - a past or current Outreachy coordinator, or
 - an employee of a current Outreachy sponsor,
   who is paid to contribute to open source as part of their job