Add employer field to MentorApproval
Some Outreachy sponsor want to know how many employees are Outreachy
mentors. In the past, we've relied on anecdotal data. Collect statistics
more formally.

Add jQuery to suggest employers based on past sponsors. Add some safe
defaults to the top of the suggestions list: "Prefer not to say", "Self
employed", and "Unemployed". Mark all mentors in past rounds as "Prefer
not to say".