Remove mentor feedback termination reason textfield
This was the only use of Wagtail's RichTextField outside of a Wagtail
page, and it apparently never worked as anything but a plain TextField.
So it needed to get fixed somehow.

But also literally nobody has ever entered anything meaningful in this
field, so Sage says let's just delete it.