Add JavaScript for country selector in initial application essays
In the initial application on the essay page, we ask applicants to fill
out what country they will be living in during the internship.

Add JavaScript to show a drop down of country names in English, along
with the country flag and the name commonly used in that country's
majority language.

Add a hidden field for the ISO 3166-1 alpha-2 country code. The hidden
field is automatically filled in by the country-select-js JavaScript
when a country is selected.

Make sure the new hidden field is allowed to be blank. That will allow
people who have JavaScript turned off to complete the form.

The country code will allow easier filter students who are ineligible
for particular rounds based on which hemisphere their country is in.

This commit adds CSS, JavaScript, and flag images from the
country-select-js repository:

https://github.com/mrmarkfrench/country-select-js

The commit that I cloned from was
2eb2b299d172278868e74fd15c92d68ff3d60239 "Merge pull request #73 from
mrmarkfrench/dependabot/npm_and_yarn/fstream-1.0.12" from June 6, 2020.

The CSS and JavaScript are under an MIT license:

https://github.com/mrmarkfrench/country-select-js#attributions

The flag images are from this repository:

https://github.com/behdad/region-flags

According the repo, the flag images are public domain images from Wikipedia:

https://github.com/behdad/region-flags/blob/gh-pages/COPYING