WIP: Remove sensitive info on initial application processing
The goal of this commit is to remove sensitive information as soon as we
process someone's initial application. This will ensure that we protect
the privacy of applicants by deleting sensitive data as soon as we are
done with it.

The sensitive information we remove is:

 - Essays about discrimination and harassment they face, and how they
   are underrepresented in the technology industry of their country
 - Gender identity
 - Race and ethnicity demographics
 - Country they will be living in during the internship
 - TODO: Reviewer comments, which may contain quotes from essays

We attempt to delete this information every time an ApplicantApproval
object goes to the APPROVED or REJECTED state.

If the ApplicantApproval goes to the PENDING state, nothing is changed.
The WITHDRAWN state is not used in the website. If an applicant asks to
withdraw their initial application, we will delete the initial
application from the website.

When an applicant finishes their initial application, we normally show
them a summary of what they've submitted. However, we need to hide
the essays, gender identity, and racial/ethnic info from applicants on
their eligibility results page.

This allows applicants to not see when we've processed their initial
application. Add a note as to why the information is missing.

Modify the view for initial application reviewers to handle this data
being deleted.

Add a test to ensure:

 - essay answers aren't visible on the eligibility results page
 - essay answers are visible to reviewers before they are deleted
 - reviewers cannot see essay answers after they have been deleted
 - TODO: check that Versions of the sensitive info are deleted
 - TODO: check that gender identity and race/ethnicity aren't visible
   and are hidden from reviewers after they have been deleted