Remove version tracking of ApplicantApproval and associated models
In the past, we used a Django plugin called Reversion to track the
change history of initial applications. This was designed to help us
track changes to the initial application essays, when we asked for essay
revisions.

Over time, the version control tracking spread to different models
related to the ApplicantApproval model.

Now that we're trying to remove sensitive information when an initial
application is processed, we need to stop tracking the version history
of those models.

Remove version history tracking of:

 - ApplicantApproval
 - BarriersToParticipation (essays and location)
 - SchoolInformation (which is an inline in the Django admin view of
   ApplicantApproval)

Add a test to ensure sensitive initial application data is not under
revision control. Remove forced revision history creation from prior
tests.