Write pseudo code to remove version history of ApplicantApproval and associated models
Originally, I wrote this as a migration that deletes all revision
history of objects associated with initial applications. However,
this migration doesn't always work. It fails when I try to run the
test suite. Also, the migration is irreversible, which is not great.

Instead, we should run something like this pseudo code on the test and
production servers before we remove the revision tracking. If you try to
remove the revision objects after you deregister the class as under
revision, the code will fail.