Allow reviewers to set essay qualities, rework essay ratings
Allow any reviewer to set a series of "essay qualities". These are
qualities that Outreachy initial application reviewers are looking for
in applicant essays.

Rework the essay rating form to only talk about whether the essay is
 - strongly compelling
 - compelling
 - weakly compelling
 - unclear
 - not compelling
 - essay answers were unrelated to questions
 - spam or trolling

Each reviewer will still be allowed to set their own opinion on whether
the essay is compelling.