Simplify logic in community_landing.html
I had to think awfully hard to understand the conditions for displaying
messages about why project details aren't visible.

First, note that both role.is_approved_applicant and role.is_reviewer
imply approved_to_see_all_project_details, as of the previous commit. So
if "not approved_to_see_all_project_details", then it's also true that
"not role.is_approved_applicant" and "not role.is_reviewer", which means
those clauses can be removed from the elif condition.

With that, both the if and the elif check the same conditions aside from
clause order and which end of the contributions period they check. But
if the contributions_open deadline hasn't passed yet, then the
contributions_close deadline hasn't passed either. So checking just the
close deadline works for both cases, and we can then check the open
deadline inside the if block.

Another way to read the result is that everyone can see project details
once the contribution period closes; but before that the explanation for
why someone might not be able to see the project details depends on
whether the contribution period has opened yet.