Allow reviewers to see project overviews
The template uses role.is_volunteer, which is any of is_organizer or
is_mentor or is_coordinator or is_reviewer, but the view only showed
projects in the first three of those cases.

Also I've pushed filtering participations into the database query
instead of checking once for each approved community, which moves the
relevant code into the view instead of the Participation model.