Add staff only page to show sponsor information
This is useful for figuring out which coordinators we need to email to
get new sponsor information. It's also useful to know before reaching
out to general funding sponsors, as they may already be a sponsor of a
community.

Add tests to ensure that only staff can view the sponsor information
page.