Keep initial application results viewable until Outreachy interns start
When Outreachy applicants receive the email that they are accepted as an
intern, the email says to check the time commitments they listed on
their initial application. That URL is the eligibility-results page.

Make sure that applicants can view their initial application until their
internship starts. That should be enough time, except for interns who
are added very late. We can add interns up to five weeks after the
standard internship start date. Ideally, those interns would also be
able to view their initial application results when they join.

However, displaying the initial application results for five weeks after
the interns start poses a problem for applicants to both the December
and May rounds. If someone applied to the December 2020 round, five
weeks puts the date into January. January is when applicants start
asking us about when initial applications will open for the May 2021
round.

So we need to have their application results not be visible starting at
the beginning of the year. Therefore, align the date to the start of the
internship.