Fix internship feedback due date.
Remove due date year in email subject.

The emails are sent within 1 week of the internship feedback being due,
so it doesn't make sense to include the year. Also, we want a shorter
subject prefix so the word "feedback" is visible in mobile email apps.

Change numeric date format from 'd' to 'j'. 'd' will show dates as 01 to
31, and 'j' will show dates as 1 to 31. We don't want the leading 0.