Clarify how mentors ask for internship extensions during feedback
The feedback form was a disaster when a mentor needed to ask for an
internship extension. This now simplifies which options mentors have.
Instead of picking a date for an extension, we ask them if they want to
extend the internship by 1 to 5 weeks (total).

We also now show them the termination options -- either terminate the
internship contract and pay the intern, or don't pay the intern.

This adds a new option for mentors: "I don't know what actions to
recommend, please advise". We typically get mentors email us saying, I
don't know what to do in this situation -- should we extend or
terminate? This gives them a way to do that via form.

All older feedback objects will have their feedback set to "I don't
know" status. That way we don't have to determine what to do with cases
where the mentor asked for extension dates that aren't in 1 week
increments.

Show the current internship dates (including custom dates clearly
marked) in the mentor feedback form. Show how many weeks extension the
intern currently has.

Keep these old fields around for JSON export, and automatically set them
when the mentor picks one of the new options:

 - request_extension
 - request_termination
 - payment_approved
 - extension_date

TODO: Display the new options on the Outreachy organizer dashboard and
feedback details views. Make sure to flag any "I don't know" statuses on
the organizer dashboard with a warning badge.