Update staff view of feedback to use new fields.
The BaseMentorFeedback now has a new "actions_requested" field that we
should use instead of the fields that individually specified whether to
pay, terminate, or extend an internship.

Modify the staff templates that display feedback status and detailed
feedback to use that new field.

Change the view so that the most important information (what action to
take) is more prominent on the feedback details page.

Display a warning color if the intern is working more than 40 hours a
week.