README: Add instructions for creating a different scenario.
It's common after working on one task to need to have the website be in
a different application phase. At that point, you want to destroy the
database and run new scenario code. However, that can be really scary to
newcomers!

Add documentation that shows them how to save a copy of the database
file, remove the current database, and recreate it.