Docs: Move week 1 blog post into Internship Guide.
Every two weeks during the Outreachy internship, interns are sent an
email with expectations for the next two weeks, and a blog prompt.

There are two motivations to move the blog prompts into the Internship
Guide. First, the weekly emails for mentors and interns are quite long.
Second, other internship programs may want to reference our blog prompts
for their own use.

Translate the email text in home/templates/home/email/internship-week-1.txt
to a new HTML section in the Internship Guide. Add a new sub-section to
the table of contents for that week's blog post prompt. Link to the blog
post prompt section in the blogging schedule.

Link the new section in the Internship Guide in the email, instead of
having the blog prompt inline in the email.

I tested that the URL displayed in the email was correct by:

1. Running the Django shell and creating an internship round that was at
   the start of week 1 of the internship:

>>> from home import scenarios
>>> scenarios.InternshipWeekScenario(week=1)

2. Creating a superuser account:

$ ./manage.py createsuperuser

3. Logging into the superuser account, going to the dashboard, and
   finding the button to preview the "Week 1" email. (It's under the
   "Email Reminders" section.)

localhost:8000/dashboard/

You should be able to see the expanded URL in the email preview. If you
do this for additional emails, please check the new URL is correct.