Docs: Add key directories for working with CSS.
I'm not a CSS expert, but someone who is organized the website CSS into
partials and sass (whatever that is?). Nicole created files in
outreachyhome/templates/sass/ during this process.

I assume if you're going to be modifying the CSS, you'd look there (or
create files elsewhere??) As I said, I'm not familiar with CSS...