Docs: Move mentor guide to be community guide for both mentors and coordinators
A lot of the information that coordinators need to know is things
mentors need to know as well. Have a "Community Guide" that includes
sections for both mentors and coordinators. Have the coordinator
sections reference mentor sections when there's overlap, or use snippets
to include the text in both sections.