WIP: Docs: community participation guide.
This documentation is for coordinators and mentors. I moved most of the
static content from the community CFP page into the community
participation guide. Added some sections as well. Some sections in the
guide are straight copied from the internship guide, and they should be
linked instead (or modified to be for a mentor/coordinator audience).

TODO: Move all content in the mentor page and mentor FAQ page into this
document. Link the community guide from the menu navigation (perhaps
renaming the "mentor" menu to the "communities" menu).