CFP: Tweak layout
Make buttons reflect which actions we want people to take:

 - Read the community guide - both mentors and coordinators need to do
   this
 - Submit a community application

I would like to have something that looks more "actiony" for finding the
right participating community to sign up to mentor for. Maybe a search
bar? Drop down? Or make the list of communities a bigger font size?
Something needs to be done to attract attention to the community list.