Add Outreachy general funding application page
Ask communities who are applying to Outreachy general funding:

1. Is your community a humanitarian open source community?
2. What humanitarian issues is your community addressing?
3. What other companies or organizations could Outreachy ask for
   additional sponsorship?

Add a skip button at the top of the page for communities that aren't
applying to the general fund.

If the sponsorship form was filled out for the first time in this
internship round, redirect to the community public info page. Otherwise
direct them to the community's mentor sign up page.