General funding form: remove CKEditor fields.
We need to make sure that Outreachy organizers never view HTML provided
by untrusted users. This can be used to inject malicious HTML and gain
access to staff-level permissions on the Outreachy website.

When a community signs up to participate for the first time, it may
provide an Outreachy general funding application. That application was
originally a CKEditor field, which allows HTML input. Change it to a
CharField to only allow plain text.