Change layout of general funding form to encourage more input.
A community signed up, wanting to use Outreachy general funding, but
didn't fill out the general funding application form. I believe they
didn't read the instruction and clicked the skip button at the top of
the page.

Remove the skip button all together, and tell communities to leave the
form blank if they are not applying to the Outreachy general fund.