CFP: Add warning for applicants who find the community CFP page.
We often get applicants who sign up to be a coordinator for a community.
I've seen an uptick of people signing up to be coordinators since the
redesign of the page.

Add a warning with bold text and a police siren emoji to tell applicants
they have the wrong page. Link to the current list of projects.

Todo: always have a link to the current list of projects from the apply
page.