Add URL for timezone conversions for the Outreachy Twitter chat.
Store a URL that encodes a timezone conversion for the start time of the
Outreachy Twitter chat.

We typically use this site:

https://www.timeanddate.com/worldclock/converter.html?p1=137&p2=179&p3=1440&p4=233&p5=125&p6=267&p7=141&p8=53&p9=367&p10=176&p11=236&p12=22

Todo: That URL has a nice API format for encoding the time and date, and
timezones we want to include. It would be good to generate that URL
programmatically for all date and times (e.g. the intern announcement).

At the same time, it's a proprietary site, so there's no guarantee it
will stick around. It would be good to research other open source
libraries that we could use to display our own timezone conversion page.