2021-05: Add blog post announcing initial application is open.
This is mostly a copy of the 2020-12 blog post. The differences are:

 - Add a button for encouraging applicants to read the Applicant Guide.
   That way they will be more likely to read the essay tips before
   submitting an initial application.

 - Add a section promoting the Twitter chat. I forgot to promote it
   in the email announcement, and we got much less interaction from
   applicants.

 - Add links to commonly asked questions, which are answered in the
   applicant guide.

 - switch the roles of the southern and northern hemisphere students in
   the section on student eligibility

The section on sensitive information data retention policy remains the
same as last round. We need to present it again for northern hemisphere
students who didn't apply last round. It can be removed in announcements
for future rounds.