Clarify what a contractor is in the initial application.
Some people are unsure what a contractor is, which can lead them to say
that they are a contractor in the initial application. Then they get
confused when they make it to the section that asks them how many hours
they typically contract per week.

Add help text that will provide more context on what a contractor is.