Fix broken link on community mentor sign up page
Fix broken URLs that reference https://www.outreachy.org/mentor/ and its
various sub-pages.

URLs that reference the base site URL should start with a "/". Otherwise
they will be interpreted as having their base URL be the current page
URL.

For example, in this HTML snippet:

<p>Please make sure you understand the <a href="mentor/#mentor">mentor requirements</a>, especially the time commitment requirement.</p>

If you're currently at the URL https://www.outreachy.org/communities/cfp/mozilla/
then the browser will append the remainder of the URL to the end of the
current page's URL. The browser will then interpret the link as:

https://www.outreachy.org/communities/cfp/mozilla/mentor/#mentor

The fix is to add a '/' at the front of the URL, so the snippet becomes:

<p>Please make sure you understand the <a href="/mentor/#mentor">mentor requirements</a>, especially the time commitment requirement.</p>

Then the browser will then interpret the link as:

https://www.outreachy.org/mozilla/mentor/#mentor

Fixes #463
Fixes #330